pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codebase-archaeologist"
version = "0.1.0"
description = "MCP server + CLI that reverse-engineers institutional knowledge from codebases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Eduard Petraeus" }]
keywords = ["mcp", "codebase", "analysis", "documentation", "archaeology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "fastmcp>=3.1.0",
    "click>=8.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[project.scripts]
archaeologist = "codebase_archaeologist.cli:cli"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
